Not sure why AITigerCompatibility would mess up on this, but I got a report of a build failure

+++ Frameworks/AIUtilities Framework/Source/AIImageAdditions.h
@@ -47,7 +47,7 @@ typedef enum {
  *
  * @result An NSData JPEG representation whose length is <= maxByteSize, or nil if one could not be made.
  */
-- (NSData *)JPEGRepresentationWithMaximumByteSize:(NSUInteger)maxByteSize;
+- (NSData *)JPEGRepresentationWithMaximumByteSize:(unsigned)maxByteSize;
 - (NSData *)PNGRepresentation;
 - (NSData *)GIFRepresentation;
 - (NSData *)BMPRepresentation;
